[src] Rename inputs_and_masks.py to transform.py
